better handle xsession: only install if dwm is installed
The issue here is that if xsession exists in the homedir, Xorg will try
to execute it by default. Given that the xsession here tries to exec
dwm, dwm must be installed and in PATH for the xsession to work.
Attempting to exec the xsession without dwm causes Xorg to crash and
exit.
Ensure that dwm is installed and in PATH before adding the xsession to
the homedir